+++ packages/vendure-plugin-webhook/src/api/request-transformer.ts
@@ -1,11 +1,18 @@
-import { Injector, RequestContext, Type, VendureEvent } from '@vendure/core';
+import {
+  Injector,
+  ProductEvent,
+  ProductVariantEvent,
+  RequestContext,
+  Type,
+  VendureEvent,
+} from '@vendure/core';
 import { Webhook } from './webhook.entity';
 
 export type TransformFn<T extends EventWithContext> = (
   event: T,
   injector: Injector,
   webhook: Webhook
-) => WebhookRequest | Promise<WebhookRequest>;
+) => WebhookRequest | Promise<WebhookRequest> | false | Promise<false>;
 
 export type EventWithContext = VendureEvent & {
   ctx: RequestContext;
@@ -19,7 +26,7 @@ export interface WebhookRequest {
 export class RequestTransformer<T extends Array<Type<EventWithContext>>> {
   readonly name: string;
   readonly supportedEvents: T;
-  readonly transform: TransformFn<EventWithContext>;
+  readonly transform: TransformFn<InstanceType<T[number]>>;
 
   constructor(
     private readonly options: {
@@ -28,7 +35,11 @@ export class RequestTransformer<T extends Array<Type<EventWithContext>>> {
        * The events that this transformer supports
        */
       supportedEvents: T;
-      transform: TransformFn<EventWithContext>;
+      /**
+       * Transform the body and/or headers for the webhook request.
+       * When returning `false`, the webhook will not be called.
+       */
+      transform: TransformFn<InstanceType<T[number]>>;
     }
   ) {
     this.name = options.name;
@@ -36,3 +47,12 @@ export class RequestTransformer<T extends Array<Type<EventWithContext>>> {
     this.transform = options.transform;
   }
 }
+
+// This only exists to test TS type compilation for this plugin, because of the complex type inference above
+new RequestTransformer({
+  name: 'Stringify Product events',
+  supportedEvents: [ProductEvent, ProductVariantEvent],
+  transform: (event: ProductEvent | ProductVariantEvent, injector) => {
+    return {};
+  },
+});